% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnophelesModel_init.R
\name{build_model_obj}
\alias{build_model_obj}
\title{incorporate vector, host and activity parameters in the entomological
model, initialize and build a model object}
\usage{
build_model_obj(vec_p, hosts_p, activity, total_pop)
}
\arguments{
\item{vec_p}{object containing vector-specific entomological parameteres
created with the function \code{def_vector_params}}

\item{hosts_p}{object containing host-specific entomological parameters
created with the function \code{def_host_params}}

\item{total_pop}{total population of humans and animals}
}
\value{
list object containing the following component:
\itemize{
\item \code{vec_params}: object with mosquito-specific entomological
parameters (use function \code{def_vector_params} to generate object)
\item \code{host_params}: object with host-specific entomological parameters
(use function \code{def_hosts_params} to generate object).
In addition to previously-defined parameters, it contains:
\itemize{
\item \code{muvA}: mosquito death rate
\item \code{alphai}: availability to mosquitoes for each host type
}
\item \code{activity}: activity patterns for mosquitoes and humans
(see function \code{def_activity_patterns} for object details)
\item \code{total_pop}: total population of humans and animals
}
}
\description{
\code{build_model_obj} This function incorporates defined
vector, host and activity parameters in the entomological model of mosquito
oviposition cycle and builds a comprehensive object containing all model
parameters. Additional parameters are calculated and included in the object.
These include mosquito death and availability rate for three different types
of hosts: humans protected by intervention, unprotected humans and animals).
The function should be applied only on list objects created by the
package functions \code{def_vector_params} and \code{def_host_params}.
Functionality using other customized objects is not guaranteed.
}
\examples{
# Retrieve the species-specific entomological parameters for An. gambiae
gambiae_vec_p = def_vector_params("Anopheles gambiae")
# Retrieve the host-specific entomological parameters for An. gambiae
gambiae_hosts_p = def_host_params("Anopheles gambiae")
# Retrieve default activity patterns for humans and mosquitoes
default_rhythms = def_activity_patterns("default_Anopheles_gambiae")
# Build model object for a population of 2000 hosts
model_p = build_model_obj(gambiae_vec_p, gambiae_hosts_p,
default_rhythms, 2000)
print(model_p)
}
\author{
Monica Golumbeanu, \email{monica.golumbeanu@swisstph.ch}
}
