% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnophelesModel.r
\name{calculate_combined_impact}
\alias{calculate_combined_impact}
\title{calculates the combined effects and impact for two interventions}
\usage{
calculate_combined_impact(
  combination_name = "combination",
  intervention1,
  intervention2,
  cov_intervention1,
  cov_intervention2,
  cov_combination = NULL,
  N_vec
)
}
\arguments{
\item{N_vec}{number of daily emerging mosquitoes}

\item{combination_name:}{name of the intervention combination}

\item{intervention1:}{intervention object generated with
\code{def_intervection_effects()}}

\item{intervention2:}{intervention object generated with
\code{def_intervection_effects()}}

\item{cov_intervention1:}{coverage of intervention 1}

\item{cov_intervention2:}{coverage of intervention 2}

\item{cov_combination:}{coverage of the combination}
}
\value{

}
\description{
\code{calculate_combined_impact} This function calculates the
proportion by which vectorial capacity is reduced when a combination of
two interventions is deployed. The user has the possibility to specify
differing coverage for each intervention and the coverage of the combination.
}
\author{
Monica Golumbeanu, \email{monica.golumbeanu@swisstph.ch}

Tom Smith, \email{thomas.smith@swisstph.ch}

Clara Champagne, \email{clara.champagne@swisstph.ch}
}
