% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnophelesModel.r
\name{calculate_impact}
\alias{calculate_impact}
\title{calculate intervention impact in terms of vectorial capacity reduction}
\usage{
calculate_impact(interventions_vec, coverage_vec, model_p, Nv0, num_ip_points)
}
\arguments{
\item{interventions_vec}{vector of intervention names}

\item{coverage_vec}{vector of coverages to consider for each intervention}

\item{model_p}{list of vector- and host-specific entomological parameters
(see description of output from function \code{calc_model_params})}

\item{Nv0}{number of mosquitoes born and entering the host-seeking state
per day}

\item{num_ip_points}{number of time points to be considered}

\item{human_behavior}{vector of human behavior before biting starts to
consider for each intervention. Can be "inside" or "outside".}
}
\value{
list object with the following attributes:
\itemize{
\item \code{vec_p}: list of vector-specific entomological parameters (see
description of output from function \code{get_vector_params})
\item \code{host_p}: list of host-specific entomological parameters (see
description of output from function \code{get_host_params})
\item \code{interventions_vec}: vector of intervention-specific entomological
parameters and impact (vectorial capacity) values
}
}
\description{
\code{calc_impact} This function calculates the
proportion by which vectorial capacity is reduced for each of a list of interventions
when deployed to a cohort of humans
}
\references{
TO DO
}
\author{
Monica Golumbeanu, \email{monica.golumbeanu@swisstph.ch}

Olivier Briët, \email{olivier.briet@swisstph.ch}

Nakul Chitnis, \email{nakul.chitnis@swisstph.ch}

Tom Smith, \email{thomas.smith@swisstph.ch}
}
