% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnophelesModel.r
\name{calculate_impact_ip}
\alias{calculate_impact_ip}
\title{calculates reduction in vectorial capacity for one interpolation point}
\usage{
calculate_impact_ip(intervention_obj, model_obj, n_vec, Ni)
}
\arguments{
\item{intervention_obj}{object with the following attributes
\itemize{
\item \code{PBi}: probability that a mosquito bites host i
\item \code{PCi}: probability that a mosquito finds a resting place after
biting a host of type i
\item \code{PDi}: probability that a mosquito survives the resting phase
after biting a host of type i
\item \code{PEi}: probability that a mosquito lays eggs and
returns to host-seeking after biting a host of type i
\item \code{Kvi}: proportion of susceptible mosquitoes that become infected
after biting any host of type i
\item \code{muvA}: mosquito death rate
\item \code{alphai}: availability to mosquitoes for host of type i
}}

\item{model_obj}{list of vector- and host-specific entomological parameters
(see description of output from function \code{calc_model_params})}

\item{Ni}{data frame with the number of protected and unprotected hosts for
various intervention coverages}

\item{N_vec}{number of daily emerging mosquitoes}
}
\value{
object with the following attributes (columns = different coverages):
#' \itemize{
\item \code{Pf_all}: probability that mosquitoes survive the feeding cycle
#' \item \code{or_all}: delayed oocyst rate
#' \item \code{sr_all}: sporozoite rate
#' \item \code{sigmai_all}: host biting rate
#' \item \code{eiri_all}: entomological innoculation rate
#' \item \code{vc_all}: vectorial capacity
#' \item \code{f_all}: duration of the feeding cycle
#' \item \code{PAmu_all}: probability that mosquitoes die during host seeking
#' \item \code{maxeig_all}: maximum eigen value of the upsilon matrix
}
}
\description{
\code{calculate_impact_ip} This function calculates the
proportion by which vectorial capacity is reduced when an intervention is
deployed to a cohort of humans at a given time point
}
\author{
Monica Golumbeanu, \email{monica.golumbeanu@swisstph.ch}

Olivier Briët, \email{olivier.briet@swisstph.ch}

Nakul Chitnis, \email{nakul.chitnis@swisstph.ch}

Tom Smith, \email{thomas.smith@swisstph.ch}
}
