% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnophelesModel.r
\name{calculate_impact_var}
\alias{calculate_impact_var}
\title{calculate intervention impact on the vectorial capacity
and assess its variation accounting for the variation of input parameters}
\usage{
calculate_impact_var(
  mosquito_species = "Anopheles gambiae",
  vec_ent_table = NULL,
  host_table = NULL,
  activity_patterns = "default_Anopheles_gambiae",
  interventions = intervention_obj_examples,
  coverage_vec = c(seq(0, 0.9, by = 0.1), 0.95, 0.99),
  total_pop = 2000,
  n_vec = 10000,
  n_time_points = 100,
  n_sample_points = 100,
  plot_result = TRUE
)
}
\arguments{
\item{mosquito_species}{string corresponding to the name of the mosquito
species to load the bionomic parameters for. This argument is case
sensitive and must be one of the species provided with the package:
\emph{Anopheles gambiae}, \emph{Anopheles albimanus}, etc.
To see the available mosquito species in th package,
you can use the function \code{list_all_species()}.
Default value is Anopheles gambiae.}

\item{vec_ent_table}{data frame with custom vector entomological values.
Must be provided if custom parameter values should be used instead of
the ones in the package database corresponding to the
provided moquito species.
The data frame should have the same structure (column names) as the database
object \code{vec_ent_param} (see Data documentation). Default value is NULL.}

\item{host_table}{data frame with custom host-specific values.
Must be provided if custom parameter values should be used instead of
the ones in the package database.
The provided data frame should have the same structure as the database
object \code{host_ent_param} (see Data documentation). Default value is NULL.}

\item{interventions}{vector of intervention names}

\item{coverage_vec}{vector of coverages to consider for each intervention}

\item{total_pop}{total population size of human and animal hosts}

\item{n_vec}{number of mosquitoes born and entering the host-seeking state
per day}

\item{n_time_points}{number of time points to be considered}

\item{n_sample_points}{number of samples to select for assessing variance}

\item{plot_result}{if TRUE, the function plots the reduction of the vectorial
capacity for each intervention in the \code{interventions} list}

\item{activity}{can be either a string corresponding to default entries:
"default_Anopheles_gambiae", "default_Anopheles_albimanus"
or a list object with the following attributes:
\itemize{
\item \code{HBI}: proportion of human biting indoors
\item \code{HBO}: propotion human biting outdoors
\item \code{humans_indoors}: proportion of humans indoor
\item \code{humans_in_bed}:proportion of humans in bed
}
These attribues can be either time series with the corresponding values at
each time point, or characters indicating an entry ID in the package
database. The function list_rhythms() can be used to retrieve the
available entries for all geographical locations.}
}
\value{
list object with the following attributes:
\itemize{
\item \code{vec_p}: list of vector-specific entomological parameters (see
description of output from function \code{get_vector_params})
\item \code{host_p}: list of host-specific entomological parameters (see
description of output from function \code{get_host_params})
\item \code{interventions_vec}: vector of intervention-specific entomological
parameters and impact (vectorial capacity) values
}
}
\description{
\code{calc_impact_var} calculates the
proportion by which vectorial capacity is reduced for each of a list of
interventions when deployed to a cohort of humans
}
\author{
Monica Golumbeanu, \email{monica.golumbeanu@swisstph.ch}
}
