% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnophelesModel_init.R
\name{def_host_params}
\alias{def_host_params}
\title{define host-specific entomological parameters}
\usage{
def_host_params(
  mosquito_species = "Anopheles gambiae",
  host_table = NULL,
  verbose = TRUE
)
}
\arguments{
\item{mosquito_species}{string corresponding to the name of the
mosquito species. Default value is "Anopheles gambiae".
To be able to use the parameters from the package database, this name must
match one of the Anopheles species provided in the package.
To see all available Anopheles species, use the function
\code{list_all_species()}.}

\item{host_table}{data frame with custom host-specific values.
Must be provided if custom parameter values should be used instead of
the ones in the package database.
The provided data frame should have the same structure (column names)
as the database object \code{host_ent_param}. Default value is NULL.}

\item{verbose}{boolean specifying whether any messages should be displayed
during the execution of the function. Default value is TRUE.}
}
\value{
list object with the following vector attributes:
\itemize{
\item \code{species_name}: name of the mosquito species
\item \code{host}: vector of hosts (should be = {"human", "animal"})
\item \code{PBi}: probability that a mosquito bites host i
\item \code{PCi}: probability that a mosquito finds a resting place after
biting a host of type i
\item \code{PDi}: probability that a mosquito survives the resting phase
after biting a host of type i
\item \code{Kvi}: proportion of susceptible mosquitoes that become infected
after biting any host of type i
}
Each vector has three elements corresponding to the probability values for:
(1) humans protected by interventions, (2) humans not protected by
interventions, and (3) animals.
In absence of interventions, the first two values of each
vector are identical.
}
\description{
\code{def_host_params} This function can be used to load
host-specific entomological parameters available in the package or to set
the entomological parameters to new custom values provided by the user.
}
\examples{
# Retrieve the host-specific entomological parameters for An. gambiae
gambiae_hosts_p = def_host_params()

# Print all the parameters
print(gambiae_hosts_p)
}
\author{
Monica Golumbeanu, \email{monica.golumbeanu@swisstph.ch}

Olivier Briët

Tom Smith

Nakul Chitnis
}
