% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnophelesModel_init.R
\name{def_vector_params}
\alias{def_vector_params}
\title{define the vector-specific entomological parameters}
\usage{
def_vector_params(
  mosquito_species = "Anopheles gambiae",
  vector_table = NULL,
  verbose = TRUE
)
}
\arguments{
\item{mosquito_species}{string corresponding to the name of the
mosquito species. Default value is "Anopheles gambiae".
To be able to use the parameters from the package database, this name must
match one of the Anopheles species provided in the package.
To see all available Anopheles species, use the function
\code{list_all_species()}.}

\item{vector_table}{data frame with custom vector entomological values.
Must be provided if custom parameter values should be used instead of
the ones in the package database.
The data frame should have the same structure (column names) as the database
object \code{vec_ent_param}. Default value is NULL.}

\item{verbose}{boolean specifying whether any messages should be displayed
during the execution of the function. Default value is TRUE.}
}
\value{
list object with the following attributes
(same as \code{vec_ent_param}):
\itemize{
\item \code{species_name}: name of the mosquito species
\item \code{M}: parous rate (proportion of host seeking mosquitoes that have
laid eggs at least once)
#' \item \code{M.sd}: standard deviation of the parous rate
\item \code{Chi}: human blood index (proportion of mosquito blood meals
derived from humans)
\item \code{A0}: sac rate (proportion of mosquitoes who laid eggs in a day)
\item \code{A0.sd}: standard deviation of the sac rate
\item \code{zeta.3}: relative availability of different non-human hosts
\item \code{td}: proportion of a day that a mosquito actively seeks a host
\item \code{tau}: time required for a mosquito that has encountered a host to
return to host seeking
\item \code{ts}: duration of the extrinsic incubation period (time required
for sporozoites to develop in mosquitoes)
\item \code{endophily:} proportion of indoor resting mosquitoes
\item \code{endophily.sd:} standard deviation of the endophily
\item \code{endophagy:} proportion of indoor feeding mosquitoes
\item \code{endophagy.sd:} standard deviation of the endophagy
}
}
\description{
\code{def_vector_params} This function can be used to load
entomological parameters for mosquito species available in the package
database or to set the entomological parameters to new custom values
provided by the user.
}
\examples{
# Retrieve the vector entomological parameters for An. gambiae
gambiae_vec_p = def_vector_params()

# Print all the parameters
print(gambiae_vec_p)

}
\author{
Monica Golumbeanu, \email{monica.golumbeanu@swisstph.ch}
}
