% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnophelesModel_db_query.R
\name{get_exposure_multiplier}
\alias{get_exposure_multiplier}
\title{calculate adjustment for averted exposure}
\usage{
get_exposure_multiplier(param_name, model_obj, intervention_type)
}
\arguments{
\item{param_name}{parameter name to adjust exposure for. Can only be one of:
"\code{alphai}", "\code{PBi}", "\code{PCi}", "\code{PDi}"}

\item{model_obj}{list of vector- and host-specific entomological parameters
(see description of output from function \code{calc_model_params})}

\item{intervention_type}{name of the type of intervention, can be only one
of: "LLINs", "IRS", "LLINsO", "Screening", "Larviciding".}
}
\value{
exposure multiplier value to be used for adjusting averted exposure
}
\description{
\code{calculate_impact_ip} This function calculates
the proportion of averted exposure to adjust for
depending on the human location importance for intervention impact
}
\references{
TO DO
}
\author{
Monica Golumbeanu, \email{monica.golumbeanu@swisstph.ch}

Olivier Briët, \email{olivier.briet@swisstph.ch}

Nakul Chitnis, \email{nakul.chitnis@swisstph.ch}

Tom Smith, \email{thomas.smith@swisstph.ch}
}
