% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnophelesModel_db_query.R
\name{get_in_out_exp}
\alias{get_in_out_exp}
\title{Calculate human exposure to biting}
\usage{
get_in_out_exp(activity_cycles, vec_p)
}
\arguments{
\item{activity_cycles}{data frame with the following columns:
\itemize{
\item \code{HBI}: proportion of human biting indoors
\item \code{HBO}: propotion human biting outdoors
\item \code{humans_indoors}: proportion of humans indoors
\item \code{humans_in_bed}:proportion of humans in bed
}}

\item{vec_p}{list of vector-specific entomological paramteres (see
description of output from function \code{def_vector_params})}
}
\value{
The function returns an object with the following attributes:
\itemize{
\item \code{Exposure_Indoor_total}: indoor exposure proportion
\item \code{Exposure_Outdoor_total}: outdoor exposure proportion
\item \code{Exposure_Indoor_whileinbed}: indoor in bed exposure proportion
\item \code{Exposure_Outdoor_whileinbed}: outdoor in bed exposure proportion
\item \code{indoor_resting}: proportion of mosquitoes that rest indoors.
}
All numbers in the output list are proportions of the overall total
exposure in the absence of interventions.
}
\description{
\code{get_in_out_exp} This function can be
used to calculate indoor and outdoor exposure to biting.
}
\references{
TO DO
}
\author{
Monica Golumbeanu, \email{monica.golumbeanu@swisstph.ch}

Olivier Briët, \email{olivier.briet@swisstph.ch}

Nakul Chitnis, \email{nakul.chitnis@swisstph.ch}

Tom Smith, \email{thomas.smith@swisstph.ch}
}
