% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnophelesModel_db_query.R
\name{get_net_decay}
\alias{get_net_decay}
\title{Get LLIN survival and holed area}
\usage{
get_net_decay(net_type, country, insecticide_type, n_ips, duration)
}
\arguments{
\item{net_type}{type of the LLIN}

\item{country}{selected country of interest for which the net_type exists}

\item{insecticide_type}{type of insecticide}

\item{n_ips}{number of time points to consider throughout the duration}

\item{duration}{considered time of the intervention, default is 3 years}
}
\value{
The function returns an object with the following attributes:
\itemize{
\item \code{time}: time point
\item \code{insecticideContent}: insecticide content of the net
\item \code{logHoles}: log holed area of the net
\item \code{survival}: survival of the net
}
}
\description{
This function can be
used to calculate LLIN survival and holed area using parameters from
the \code{AnophelesModel} database for a given net type and country.
To list all available net types and corresponding countries,
use the function get_net_types().
}
\examples{
net_surv = get_net_decay(net_type = "DuraNet",
country = "Kenya", insecticide_type = "Default", n_ips = 100, duration = 3)
print(net_surv)

}
\author{
Monica Golumbeanu, \email{monica.golumbeanu@swisstph.ch}

Olivier Briët, \email{olivier.briet@swisstph.ch}

Nakul Chitnis, \email{nakul.chitnis@swisstph.ch}

Tom Smith, \email{thomas.smith@swisstph.ch}
}
