% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnophelesModel_db_query.R
\name{get_net_insecticides}
\alias{get_net_insecticides}
\title{Get net insecticide types}
\usage{
get_net_insecticides()
}
\value{
The function returns a name list of net insecticide types.
}
\description{
This function displays a table of the
available LLIN insecticide types
available within the \code{AnophelesModel} database.
}
\examples{
net_insecticide_types = get_net_insecticides()
print(net_insecticide_types)

}
\author{
Monica Golumbeanu, \email{monica.golumbeanu@swisstph.ch}
}
