% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnophelesModel_db_query.R
\name{get_net_types}
\alias{get_net_types}
\title{Get net types}
\usage{
get_net_types()
}
\value{
The function returns a table of net types and corresponding countries
}
\description{
This function displays a table of the
available LLIN types and the countries for which durability profiles
are available within the \code{AnophelesModel} database
}
\examples{
net_types = get_net_types()
print(net_types)

}
\author{
Monica Golumbeanu, \email{monica.golumbeanu@swisstph.ch}
}
