% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnophelesModel_xml.R
\name{get_OM_GVI_snippet}
\alias{get_OM_GVI_snippet}
\title{creates an xml snippet to be included in an OpenMalaria scenario
for the GVI intervention}
\usage{
get_OM_GVI_snippet(species, intervention_impact, num_points, plot_f = TRUE)
}
\arguments{
\item{intervention_impact}{object created by the AnophelesModel main function}
}
\value{
TO DO
}
\description{
\code{get_OM_GVI_snippet} This function can be used to
create the relevant xml tags for the <GVI> component of the
scenario xml file required as input for running the OpenMalaria
agent-based simulation model.
Precisely, it creates the <mosq> element containing
parameters describing the mosquito life cycle, as well as the <nonHumanHosts>
element with parameters specific to animal hosts.
}
\references{
TO DO
}
\author{
Monica Golumbeanu, \email{monica.golumbeanu@swisstph.ch}

Olivier Briët, \email{olivier.briet@swisstph.ch}

Nakul Chitnis, \email{nakul.chitnis@swisstph.ch}

Tom Smith, \email{thomas.smith@swisstph.ch}
}
