% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnophelesModel_data.R
\docType{data}
\name{host_ent_param}
\alias{host_ent_param}
\title{Host-specific estimated parameters for the entomological model of
malaria transmission}
\format{
A \code{data frame} with 5 columns consisting of:
\itemize{
\item \code{species name:} name of the mosquito species
\item \code{probability:} probabilities that a mosquito completes the next
stage of the oviposition cycle in one night, conditional on having reached
that stage. Can only be one of the following:
PBi(probability that a mosquito bites host of type i), PCi(probability that a
mosquito finds a resting place after biting a host of type i), PDi
(probability that a mosquito survives the resting phase after
biting a host of type i), PEi(probability that a mosquito lays eggs and
returns to host-seeking after biting a host of type i), Kvi(proportion of
susceptible mosquitoes that become infected after biting a host of type i)
\item \code{host_type_1:} value of the probability for host type 1 (humans
protected by the intervention)
\item \code{host_type_2:} value of the probability for host type 2 (humans
not protected by any intervention)
\item \code{host_type_3:} value of the probability for host type 3 (animals)
}
}
\usage{
host_ent_param
}
\value{
host-specific entomological parameters
}
\description{
This data object contains for each of the three host types (humans
protected by interventions, non-protected humans, and animals) estimated
parameters for the entomological model of malaria transmission.
}
\examples{
# Print the first lines of the data frame object
head(host_ent_param)

}
\references{
TO DO
}
\author{
Monica Golumbeanu, \email{monica.golumbeanu@swisstph.ch}

Olivier Briët, \email{olivier.briet@swisstph.ch}

Nakul Chitnis, \email{nakul.chitnis@swisstph.ch}

Tom Smith, \email{thomas.smith@swisstph.ch}
}
\keyword{datasets}
