% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnophelesModel_data.R
\docType{data}
\name{interventions_param}
\alias{interventions_param}
\title{General description of IRS and LLINs interventions parameterizations}
\format{
A \code{data frame} with 7 columns consisting of:
\itemize{
\item \code{Parameterisation:} ID of the parameterization
\item \code{Intervention:} name identified of the intervention, can only take
 value "IRS", "LLINs" or "House screening"
\item \code{Active agent:} insecticide name
\item \code{Species:} name of the mosquito species
\item \code{Reference:} publication reference
\item \code{Method:} fitting method
\item \code{Duration:} duration until intervention renewal
}
}
\usage{
interventions_param
}
\value{
general intervention description
}
\description{
This data object contains a general description of all available
IRS and LLIN intervention parameterizations.
}
\examples{
# Print the first lines of the data frame object
head(interventions_param)

}
\author{
Monica Golumbeanu, \email{monica.golumbeanu@swisstph.ch}

Olivier Briët,

Nakul Chitnis, \email{nakul.chitnis@swisstph.ch}

Tom Smith, \email{thomas.smith@unibas.ch}
}
\keyword{datasets}
