% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnophelesModel_db_query.R
\name{list_activity}
\alias{list_activity}
\title{Extract activity patterns of humans or mosquitoes}
\usage{
list_activity(
  species = "Anopheles gambiae",
  sampling = "HBI",
  country = NULL,
  site = NULL
)
}
\arguments{
\item{species}{string corresponding to the name of the species to extract the
activity patterns for. For human activity patterns, this parameters should be
"Homo Sapiens". For mosquito activity patterns, this name must match one of
the Anopheles species provided in the package.
The default value is "Anopheles gambiae".
To see all available Anopheles species, use the function list_all_species().}

\item{sampling}{can only be one of the following strings:
\itemize{
\item \code{IND}: proportion of humans indoors; only relevant for humans
(species parameter is "Homo Sapiens").
\item \code{BED}: proportion of humans in bed; only relevant for humans
(species parameter is "Homo Sapiens").
\item \code{HBI}: proportion of mosquito biting of humans occurring indoors;
only relevant for mosquito species.
\item \code{HBO}: proportion of mosquito biting of humans occurring outdoors;
only relevant for mosquito species.
#' \item \code{ABO}: proportion of mosquito biting of animals occurring
outdoors; only relevant for mosquito species.
#' \item \code{HB}: proportion of mosquito biting of humans
(indoors and outdoors); only relevant for mosquito species.
The default value is "HBI".
}}

\item{country}{name of the country. Default value is NULL. If the country is
not provided for the provided species and
sampling code, then data for all available countries will be displayed.}

\item{site}{name of the geographical location within the country.
Default value is NULL. If this parameter is
not provided for the provided species and
sampling code, then data for all available sites and countries will
be displayed).}
}
\value{
data frame with the following columns:
\itemize{
\item \code{ID}: id of the database entry
\item \code{species}: name of the considered species
\item \code{sampling}: sampling type
\item \code{country}: country where measurements were taken
\item \code{site}: geographical site where measurements were taken
\item \code{variable}: time points; these are chronologically ordered in the
output data frame
\item \code{value}: value of the sampling at each time point
}
}
\description{
This function returns the activity patterns
of humans or mosquitoes available in the package database based on a list of
provided criteria.
}
\examples{
# List the proportion of human biting indoors in the Rachuonyo region, Kenya.
humans_indoors = list_activity(species = "Homo sapiens", sampling = "IND",
site="Rachuonyo")
print(humans_indoors)

}
\author{
Monica Golumbeanu, \email{monica.golumbeanu@swisstph.ch}
}
