% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnophelesModel_db_query.R
\name{list_all_species}
\alias{list_all_species}
\title{Print all the species whose entomological parameters are provided in
the package database.}
\usage{
list_all_species()
}
\value{
The function prints all the species names and also
returns a vector containing them.
}
\description{
This function can be used to list
all mosquito species whose entomological parameters are available in the
package.
}
\examples{
species_vec = list_all_species()
print(species_vec)

}
\author{
Monica Golumbeanu, \email{monica.golumbeanu@swisstph.ch}
}
