% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnophelesModel_db_query.R
\name{list_available_interventions}
\alias{list_available_interventions}
\title{Print all the interventions available in the package}
\usage{
list_available_interventions()
}
\value{
The function prints all the interventions and also
returns a vector containing them.
}
\description{
This function can be
used to list all interventions whose impact
models parameterisations are available in the package.
}
\examples{
interv_vec = list_available_interventions()
print(interv_vec)

}
\author{
Monica Golumbeanu, \email{monica.golumbeanu@swisstph.ch}
}
