% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnophelesModel_db_query.R
\name{list_intervention_models}
\alias{list_intervention_models}
\title{Print all the interventions parameterisations available in the package}
\usage{
list_intervention_models()
}
\value{
The function prints all the interventions models and also
returns a vector containing them.
}
\description{
This function can be used to list
all interventions models parameterisations available in the package.
}
\examples{
interv_models = list_intervention_models()
print(interv_models)

}
\author{
Monica Golumbeanu, \email{monica.golumbeanu@swisstph.ch}
}
