% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnophelesModel_plot.R
\name{plot_impact_species}
\alias{plot_impact_species}
\title{produce various impact figures per species}
\usage{
plot_impact_species(impacts_obj, plot_type)
}
\arguments{
\item{mosquito_species}{string corresponding to the name of the mosquito
species that the parameters should be loaded. Argument is case sensitive and
Must be one of the species provided with the package: Gambiae, Albimanus, ...}
}
\value{
plot object
}
\description{
\code{plot_impact_species}
}
\references{
TO DO
}
\author{
Monica Golumbeanu, \email{monica.golumbeanu@swisstph.ch}

Olivier Briët, \email{olivier.briet@swisstph.ch}

Nakul Chitnis, \email{nakul.chitnis@swisstph.ch}

Tom Smith, \email{thomas.smith@swisstph.ch}
}
