% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnophelesModel_plot.R
\name{plot_impact_var}
\alias{plot_impact_var}
\title{produce various impact figures per species}
\usage{
plot_impact_var(species_name, plot_df)
}
\arguments{
\item{species_name}{name of the mosquito species}

\item{plot_df}{plot data frame}
}
\value{
plot object
}
\description{
\code{plot_impact_var}
}
\author{
Monica Golumbeanu, \email{monica.golumbeanu@swisstph.ch}
}
